board API에서 vaildation 분리
